KafkaController, StatusService, new endpoints in StatusController

+++ src/main/java/com/softserve/kickscooter/kickscootervehicle/management/dto/ScooterStatusDto.java
@@ -8,8 +8,9 @@ import java.util.UUID;
 
 @Data
 public class ScooterStatusDto {
+
     UUID id;
     ScooterStatus status;
     Point gpsPoint;
-    Integer battery;
+    Short battery;
 }
